refactored system message build components for easier discovery with autocomplete
