+++ openai_function_call/dsl/__init__.py
@@ -1,5 +1,5 @@
 from .completion import ChatCompletion
-from .multitask import MultiTask
 from .messages import *
+from .multitask import MultiTask
 
 __all__ = ["ChatCompletion", "MultiTask", "messages"]
