+++ openai_function_call/dsl/messages/base.py
@@ -1,0 +1,28 @@
+from enum import Enum, auto
+from typing import List, Optional
+
+from pydantic import Field
+from pydantic.dataclasses import dataclass
+
+
+class MessageRole(Enum):
+    USER = auto()
+    SYSTEM = auto()
+    ASSISTANT = auto()
+
+
+@dataclass
+class Message:
+    content: str = Field(default=None, repr=True)
+    role: MessageRole = Field(default=MessageRole.USER, repr=False)
+    name: Optional[str] = Field(default=None)
+
+    def dict(self):
+        assert self.content is not None, "Content must be set!"
+        obj = {
+            "role": self.role.name.lower(),
+            "content": self.content,
+        }
+        if self.name and self.role == MessageRole.USER:
+            obj["name"] = self.name
+        return obj
